fix: numpy.float32 JSON 序列化崩溃 + README 数据治理系统更新
- 修复 pgvector + numpy 环境下 round() 保留 float32 类型导致 json.dumps 崩溃
- _cosine_distance / retrieval similarity / matching phase 显式 float() 归一化
- _push_thinking 新增 _SafeJsonEncoder 兜底安全网
- README 新增数据治理系统章节（智能路由/知识沉淀/概念网络/内容治理）
- 顶层 mermaid 流程图更新为知识闭环
- 新增 docs/bugfix/numpy-float32-json-serialization.md 记录

+++ backend/app/services/retrieval_service.py
@@ -117,7 +117,7 @@ async def _pgvector_search(
             chunk_id=row.chunk_id,
             doc_name=row.doc_name,
             content=row.content,
-            similarity=round(row.similarity, 4),
+            similarity=float(round(row.similarity, 4)),
             page=row.page,
             result_type=row.result_type,
         )
